+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.0.0] - 2018-11-18
+
+### Refactor
+- Build becomes Child
+- Skill becomes Item
+- Forked into skill tree project
+
 ## [1.0.0] - 2018-07-08
 
 ### Refactor